+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+- 2024-11-19:
+  - updated packages.
+  - refactored API code to remove `Startup` class.
 - 2024-11-17: replaced Swagger with Scalar (open `scalar/v1`) leveraging .NET 9 OpenAPI components:
   - remove package `Swashbuckle.AspNetCore`.
   - add packages `Microsoft.AspNetCore.OpenApi` and `Scalar.AspNetCore`.